feat(userAnalytic): creation of the user analytics file
